refactor: wip splitting title generation and executions from chat-hub service
