+++ packages/cli/src/modules/chat-hub/chat-message.repository.ts
@@ -1,5 +1,10 @@
-import type { ChatHubMessageStatus, ChatMessageId, ChatSessionId } from '@n8n/api-types';
-import { withTransaction } from '@n8n/db';
+import type {
+	ChatHubConversationModel,
+	ChatHubMessageStatus,
+	ChatMessageId,
+	ChatSessionId,
+} from '@n8n/api-types';
+import { User, withTransaction } from '@n8n/db';
 import { Service } from '@n8n/di';
 import { DataSource, EntityManager, Repository } from '@n8n/typeorm';
 import { QueryDeepPartialEntity } from '@n8n/typeorm/query-builder/QueryPartialEntity';
@@ -7,6 +12,7 @@ import { UnexpectedError, type IBinaryData } from 'n8n-workflow';
 
 import { ChatHubMessage } from './chat-hub-message.entity';
 import { ChatHubSessionRepository } from './chat-session.repository';
+import { EditMessagePayload, HumanMessagePayload } from './chat-hub.types';
 
 @Service()
 export class ChatHubMessageRepository extends Repository<ChatHubMessage> {
@@ -36,6 +42,66 @@ export class ChatHubMessageRepository extends Repository<ChatHubMessage> {
 				{ lastMessageAt: new Date() },
 				em,
 			);
+		});
+	}
+
+	async createHumanMessage(
+		payload: HumanMessagePayload | EditMessagePayload,
+		attachments: IBinaryData[],
+		user: User,
+		previousMessageId: ChatMessageId | null,
+		model: ChatHubConversationModel,
+		revisionOfMessageId?: ChatMessageId,
+		trx?: EntityManager,
+	) {
+		await this.createChatMessage(
+			{
+				id: payload.messageId,
+				sessionId: payload.sessionId,
+				type: 'human',
+				status: 'success',
+				content: payload.message,
+				previousMessageId,
+				revisionOfMessageId,
+				name: user.firstName || 'User',
+				attachments,
+				...model,
+			},
+			trx,
+		);
+	}
+
+	async createAIMessage({
+		id,
+		sessionId,
+		executionId,
+		previousMessageId,
+		content,
+		model,
+		retryOfMessageId,
+		status,
+	}: {
+		id: ChatMessageId;
+		sessionId: ChatSessionId;
+		previousMessageId: ChatMessageId | null;
+		content: string;
+		model: ChatHubConversationModel;
+		executionId?: string;
+		retryOfMessageId: ChatMessageId | null;
+		editOfMessageId?: ChatMessageId;
+		status?: ChatHubMessageStatus;
+	}) {
+		await this.createChatMessage({
+			id,
+			sessionId,
+			previousMessageId,
+			executionId: executionId ? parseInt(executionId, 10) : null,
+			type: 'ai',
+			name: 'AI',
+			status,
+			content,
+			retryOfMessageId,
+			...model,
 		});
 	}
